refactor: mpeg parser into files
