+++ src/demo/src/isobmff/mpeg/Header.js
@@ -20,72 +20,7 @@ TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR TH
 SOFTWARE.
 */
 
-export default class MP3Parser {
-  constructor() {
-    this._headers = new Map();
-  }
-
-  _getHeader(buffer) {
-    if (buffer.length >= 4) {
-      const key = String.fromCharCode.apply(null, buffer.subarray(0, 4));
-
-      if (this._headers.has(key)) {
-        console.log("getting header from cache");
-        return this._headers.get(key);
-      } else {
-        const headerValues = Header._buildHeader(buffer);
-        if (headerValues) {
-          console.log(headerValues);
-
-          const header = new Header(headerValues);
-          this._headers.set(key, header);
-          return header;
-        }
-      }
-    }
-  }
-
-  /**
-   * @description Finds and returns an MPEG frame. Frame will be null if no frame was found in the data
-   * @param {Uint8Array} data MPEG data
-   * @param {number} offset Offset where frame should be
-   * @returns {object} Object containing the actual offset and frame.
-   */
-  readFrame(data, offset = 0) {
-    let header = this._getHeader(data.subarray(offset));
-
-    while (!header && offset + 4 < data.length) {
-      offset++;
-      header = this._getHeader(data.subarray(offset));
-    }
-
-    return {
-      offset,
-      frame: header ? new Frame(data.subarray(offset), header) : null,
-    };
-  }
-}
-
-class Frame {
-  constructor(buffer, header) {
-    this._header = header;
-    this._data = buffer.subarray(0, header.frameByteLength);
-  }
-
-  get header() {
-    return this._header;
-  }
-
-  get data() {
-    return this._data;
-  }
-
-  get isComplete() {
-    return !(this._header.frameByteLength - this._data.length);
-  }
-}
-
-class Header {
+export default class Header {
   static v1l1Bitrates = {
     "0000": "free",
     "0001": 32,
